# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 同城O2O配送系统 API
  version: 1.0.0
  description: |
    统一的 REST API 规格，响应格式严格遵循 PRD《接口规范与状态机.md》：
    - 成功：`{ code: 0, message, data }`
    - 分页：`data.list[]` + `data.pagination{page,pageSize,total,totalPages}`
    - 错误：`{ code, message, data: null, errors? }`

    本骨架只覆盖"阶段02 数据库设计与基础服务"已交付的接口；
    阶段03 起的业务接口在对应服务域开发时补充。

servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发
  - url: https://api.example.com/api/v1
    description: 生产环境（占位）

tags:
  - name: Common
    description: 公共能力（健康检查、短信、文件上传）
  - name: Auth-User
    description: 用户认证（微信登录、绑定手机号）
  - name: Auth-Merchant
    description: 商家认证（注册、登录）
  - name: Auth-Rider
    description: 骑手认证（注册、登录）
  - name: Auth-Admin
    description: 管理员认证
  - name: Auth-Common
    description: Token 公共操作（刷新、登出）

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Success:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          description: 业务数据，具体结构见每个接口
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 10001
        message:
          type: string
          example: 参数校验失败
        data:
          nullable: true
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          example: 7200
    AdminInfo:
      type: object
      required: [id, username, roleCode, permissions]
      properties:
        id:
          type: integer
        username:
          type: string
        realName:
          type: string
        avatar:
          type: string
        roleName:
          type: string
        roleCode:
          type: string
        permissions:
          type: array
          items:
            type: string

  responses:
    OK:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未登录或 Token 过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: 请求频率超限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

paths:
  /common/health:
    get:
      tags: [Common]
      summary: 健康检查
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '503':
          description: 依赖不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /common/sms/send-code:
    post:
      tags: [Common]
      summary: 发送短信验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  pattern: '^1[3-9]\d{9}$'
                scene:
                  type: string
                  enum: [LOGIN, REGISTER, BIND_PHONE, RESET_PASSWORD]
                  default: LOGIN
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  /common/upload/image:
    post:
      tags: [Common]
      summary: 上传单张图片
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/user/wx-login:
    post:
      tags: [Auth-User]
      summary: 微信一键登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                encryptedData:
                  type: string
                iv:
                  type: string
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /auth/user/bind-phone:
    post:
      tags: [Auth-User]
      summary: 绑定手机号
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone:
                  type: string
                code:
                  type: string
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: 手机号已被占用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/user/profile:
    put:
      tags: [Auth-User]
      summary: 更新用户资料
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                avatar:
                  type: string
                gender:
                  type: integer
                  enum: [0, 1, 2]
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/merchant/send-code:
    post:
      tags: [Auth-Merchant]
      summary: 商家发送验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
                scene: { type: string, default: LOGIN }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  /auth/merchant/register:
    post:
      tags: [Auth-Merchant]
      summary: 商家注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code, password, name, contactName]
              properties:
                phone: { type: string }
                code: { type: string }
                password: { type: string }
                name: { type: string }
                contactName: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: 手机号已注册
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/merchant/login:
    post:
      tags: [Auth-Merchant]
      summary: 商家验证码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: 商家账号不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/merchant/login-pwd:
    post:
      tags: [Auth-Merchant]
      summary: 商家密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone: { type: string }
                password: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: 密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/rider/send-code:
    post:
      tags: [Auth-Rider]
      summary: 骑手发送验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
                scene: { type: string, default: LOGIN }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /auth/rider/register:
    post:
      tags: [Auth-Rider]
      summary: 骑手注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code, password, name]
              properties:
                phone: { type: string }
                code: { type: string }
                password: { type: string }
                name: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: 手机号已注册
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/rider/login:
    post:
      tags: [Auth-Rider]
      summary: 骑手验证码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string }
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /auth/admin/login:
    post:
      tags: [Auth-Admin]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/TokenPair'
                          - type: object
                            properties:
                              adminInfo:
                                $ref: '#/components/schemas/AdminInfo'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/admin/info:
    get:
      tags: [Auth-Admin]
      summary: 获取当前管理员信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 当前管理员信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminInfo'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /auth/refresh-token:
    post:
      tags: [Auth-Common]
      summary: 刷新 Access Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: 新的 Token 对
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenPair'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /auth/logout:
    post:
      tags: [Auth-Common]
      summary: 登出（将当前 Token 加入黑名单）
      security:
        - BearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/OK' }
        '401': { $ref: '#/components/responses/Unauthorized' }